safari-600.7-branch/LayoutTests/inspector-obsolete/console/console-command-clear.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/console-test.js"></script>
<script>

function log() {
    // Fill console.
    console.log("one");
    console.log("two");
    console.log("three");
}

log();

function test()
{

    InspectorTest.addResult("=== Before clear ==="); 
    InspectorTest.dumpConsoleMessages();

    function callback()
    {
        InspectorTest.addResult("=== After clear ==="); 
        InspectorTest.dumpConsoleMessages();
        InspectorTest.completeTest();
    }
    InspectorTest.evaluateInConsole("clear()", callback);
}

</script>
</head>

<body onload="runTest()">
<p>
    Tests that console is cleared upon clear() eval in console.
</p>

</body>
</html>
